+++ src/lib/config.ts
@@ -61,9 +61,9 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 
 export const SDK_METADATA = {
   language: "typescript",
-  openapiDocVersion: "4.15.0-alpha.1762501729624-b5c2d147",
-  sdkVersion: "0.3.0-alpha.2",
+  openapiDocVersion: "4.15.0-alpha.1762556927854-f0fbba4e",
+  sdkVersion: "0.3.0-alpha.3",
   genVersion: "2.743.2",
   userAgent:
-    "speakeasy-sdk/typescript 0.3.0-alpha.2 2.743.2 4.15.0-alpha.1762501729624-b5c2d147 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.3.0-alpha.3 2.743.2 4.15.0-alpha.1762556927854-f0fbba4e cribl-control-plane",
 } as const;